Added code to access tokens
