+++ project/spotify.py
@@ -14,16 +14,3 @@ def add_to_playlist(playlist_id, songs):
 def reset_and_add_to_playlist(playlist_id, songs):
     sp = spotipy.Spotify(auth='BQDpwmu5IqF4KopCSn11J-DdvSXJxz2-rWptMuurQ_lWk2p5n6ES6izN_0G7IRz_sJI-3uafxO-h-CI20WqRwEYkeid85zQAniy_w1bBRsdwPp9L7L1jekddP6ToE6TZ9xUyZheFZnXZyzV8HZjGS-lX0v_VnyQjc3dYzskizS5CqwQIiCJsfZnC4yEXam-0eWN9IjkmBZcIwJUT9mTVlmIDqxmiDD3l33UuciHNZqDg')
     sp.user_playlist_replace_tracks(cred.client_username, playlist_id, songs)
-
-def get_auth_object():
-    #Need to more access to user data? Add more scopes here!
-    SCOPE = 'playlist-read-private user-library-read playlist-modify-public playlist-modify-private'
-    CACHE = '.spotipyoauthcache'
-    SPOTIPY_REDIRECT_URI = "http://127.0.0.1:5000/auth/"
-    #TODO determine where to redirect after authorization by putting URI in SPOTIPY_REDIRECT_URI
-    sp_oauth = oauth2.SpotifyOAuth(cred.client_id, cred.client_secret,SPOTIPY_REDIRECT_URI,scope=SCOPE,cache_path=CACHE)
-
-    #print(sp_oauth.get_cached_token())
-    #print(sp_oauth.is_token_expired())
-
-    return sp_oauth
